!status completed
v2024.5.6, | Totals: 21 channels, 1 users, 3 scriptures, 0 Gospel Proclamations! | Current Metrics: 00d 0h 3m uptime, 2 messages parsed (0ms avg), 7 responses (1ms avg) (0ms)

+++ lib/twitch/src/common/message_data.rs
@@ -86,11 +86,11 @@ impl MessageData {
         types
     }
 
-    pub fn complete(&self) -> Result<u128, &'static str> {
+    pub fn complete(&self) -> Result<u64, &'static str> {
         SystemTime::now()
             .duration_since(self.received)
             .map_err(|_| "Time went backwards")
-            .map(|dur| dur.as_millis()) // Convert Duration to milliseconds
+            .map(|dur| dur.as_millis() as u64)
     }
 
     pub fn new(raw: &str) -> Option<Self> {
